+++ lib/dependencies.sh
@@ -27,17 +27,10 @@ run_if_present() {
 
 yarn_node_modules() {
   local build_dir=${1:-}
+
   echo "Installing node modules (yarn.lock)"
   cd "$build_dir"
-  yarn install --pure-lockfile --ignore-engines --cache-folder $build_dir/.cache/yarn 2>&1
-  # according to docs: "Verifies that versions of the package dependencies in the current project’s package.json matches that of yarn’s lock file."
-  # however, appears to also check for the presence of deps in node_modules, so must be run after install
-  if $(yarn check 1>/dev/null); then
-    echo "yarn.lock and package.json match"
-  else
-    echo "yarn.lock is outdated"
-    warning "yarn.lock is outdated." "run \`yarn install\`, commit the updated \`yarn.lock\`, and redeploy"
-  fi
+  yarn install --pure-lockfile --ignore-engines 2>&1
 }
 
 npm_node_modules() {
@@ -51,7 +44,7 @@ npm_node_modules() {
     else
       echo "Installing node modules (package.json)"
     fi
-    npm install --unsafe-perm --userconfig $build_dir/.npmrc --cache $build_dir/.npm 2>&1
+    npm install --unsafe-perm --userconfig $build_dir/.npmrc 2>&1
   else
     echo "Skipping (no package.json)"
   fi